refactor(oauth): rename vault-mcp/auth/ -> oauth/ to disambiguate from src/auth.ts
The folder is specifically the OAuth 2.1 server (provider, routes, consent), and
"auth/" collided conceptually with the shared src/auth.ts token utilities
(safeEqual, parseBearer). oauth/ is precise. git mv preserves history; only
server.ts's two import specifiers change (intra-folder imports keep the same
relative depth). AGENTS.md structure tree + layering note updated to match.
