+++ src/vault-mcp/oauth/__tests__/consent-page.test.ts
