+++ src/vault-mcp/oauth/__tests__/oauth-provider.test.ts
