+++ src/vault-mcp/oauth/__tests__/oauth-routes.test.ts
