+++ src/vault-mcp/oauth/consent-page.ts
