+++ src/vault-mcp/oauth/oauth-provider.ts
