+++ src/vault-mcp/oauth/oauth-routes.ts
