+++ src/vault-mcp/server.ts
@@ -7,8 +7,8 @@ import { fileURLToPath } from "node:url"
 import { createSearchIndex } from "./search/search-index.js"
 import { createMemoryStore } from "./vault-operations/memory-store.js"
 import { startFileWatcher } from "./search/file-watcher.js"
-import { createOAuthProvider } from "./auth/oauth-provider.js"
-import { createOAuthRoutes } from "./auth/oauth-routes.js"
+import { createOAuthProvider } from "./oauth/oauth-provider.js"
+import { createOAuthRoutes } from "./oauth/oauth-routes.js"
 import { createMcpRouter } from "./mcp-core/mcp-router.js"
 import { loadConfig } from "./config.js"
 import { logger } from "../logger.js"
